k8s-api/requirements.txt
Flask
flask-restx
psycopg2-binary
python-json-logger
sqlalchemy
pyyaml
flask-cors